i#5725: Set consistent x86 vendor decoding default (#6438)
If we initialize the proc_get_vendor() to VENDOR_UNKNOWN we get
contradictory decoding results for x86 opcodes that vary between
VENDOR_AMD and VENDOR_INTEL (because some of our decoding code checks
one and some checks the other), so we pick one (VENDOR_INTEL) for a
stable self-consistent default. We set this statically and we override
the underlying hardware in proc_init(). The user can use
proc_set_vendor() (after an explicit init) to override if desired.

Fixes #5725