i#2062: Support filtering of non-module blocks
Adds support for instrumenting non-module code instr-by-instr instead of
whole-block-at-a-time. This is required for the L0 filter mode which
instruments individual instructions rather than one pc entry for the
whole block.

Adds a new variant of the tool.drcacheoff.gencode test that runs with
the L0_filter enabled. Modifies the test to add a sequence of instrs
that produces an error in raw2trace due to an apparently out-of-block
memref.

This changes how we store the location of non-module code in trace PC
entries. Previously all PC_MODOFFS_BITS (33)
bits were used to store the gencode block id, with no bits assigned to
offset into the block. Now we use PC_BLOCKOFFS_BITS
(equal to PC_INSTR_COUNT_BITS, 12) bits for offset into the module
 and PC_BLOCKIDX_BITS (21) bits for block idx.