i#6471 sched idle: Add idle time from blocking syscalls
Adds a longer waiting period on blocking syscalls using either
provided output time or for instruction-based quanta a count of queue
selections before a blocked input is actually selected.  Since the
scheduler does not have timer interrupts or regular points of control
and relies on its user calling it, idle inputs are kept on the ready
queue and are checked for becoming unblocked when the ready queue is
queried.

The wait duration is set based on the "wait time factor" which is the
syscall latency divided by the context switch threshold multipled by a
user-provided "block_time_scale" option which can be used to scale up
or down the durations.

The wait duration is erased on a direct switch to an input.

Adds a new replay record type to represent idle time on replay.

Augments the unit tests to include blocking high-latency syscalls to
test the new feature in various sub-tests.

Issue: #6471