i#5505 kernel tracing: Add syscall instr encodings
Adds encodings for kernel system call instructions to the trace in raw2trace.
Kernel system call traces are decoded using libipt which also provides the
instruction encodings. We write these encodings to a new buffer which is re-used
for all dynamic instances of that instr.

Adds a new drcachesim option to ignore failures in decoding instructions. This
is currently required because the kernel traces have opcodes that are not yet
recognized. Such decode failures are not critical because their impact is
limited (not counting unsupported instrs in the opcode_mix tool, or not showing
unsupported instrs in the view tool) unlike user-space instrs where missing
decoder support may lead to drreg issues etc. Uses the new option in opcode_mix
and the view tool.

Adds support in the syscall_mix tool to report the counts of each system call's
trace also.

Issue: #5505