i#5505 kernel tracing: Support clac and stac in DR decoder (#6484)
Adds encoding-decoding support for clac and stac which are observed in
kernel traces.

decode_fast does not need any change and already should get the 3-byte
size correctly.

Adds EFLAGS_READ_AC and EFLAGS_WRITE_AC to track reads/writes to the
Alignment Check flag.
Currently we try to avoid a major binary compatibility break by adding
them to the end. This makes
some code more complicated. Added a TODO with #6485 to re-number when
we're about to do the
next major release.

Issue: #5505
Fixes: #2103