i#5365 Fix some tests which fail when build with -O3 (#6492)
- client.drsyms-test and client.drwrap-test-detach: The tests expect to
observe a certain function call a certain sub-function but it doesn't
happen when built with optimisation on because the sub-function gets
inlined. This is fixed by marking the sub-functions as NOINLINE.

- client.drx-scattergather and client.drx-scattergather-bbdup: The test
clients used with these tests count the number of scatter/gather
instructions that are expanded and print the number at the end of the
test, which gets checked against a reference value. Building the test
app with -O3 causes some code to be auto vectorized so there are
additional scatter/gather instructions which throws off the count. I
removed this tests from the sve_tests list so it won't be built with
-O3.

Issue: #5365