i#5036 A64 scatter/gather, part 9: State restore
Adds support for restoring register app state when a fault is triggered
by one of the instructions in a scatter/gather expansion sequence
emitted by drx_expand_scatter_gather().

Also extends the existing scatter/gather expansion tests to check the value
of the FFR register, and run the test instructions with different elements
rigged to fault so we can test the state restoration behaviour.

Support for the first-faulting and non-faulting variants (ldff1*, ldnf1*)
will be added in future commits.

Issue: #5036