i#5694 core-sharded: Add core-serial support
Adds a new scheduler option single_lockstep_output which multiplexes
the virtual core output streams onto a single global stream.  This is
simple to implement as the existing scheduler_t::stream_t class
already multiplexes inputs onto an output.

Hooks up the drcachesim launcher -core_serial option to this new
scheduler mode.

Updates the schedule_stats, basic_counts, and cache_simulator tools to
support core_serial.  For cache_simulator, the existing thread-to-core
mapping code for round-robin and for -cpu_scheduling is kept for when
in thread-sharded mode; in core-sharded mode, the scheduler's cpuid is
mapped to a core index.

Adds a core_serial test of schedule_stats and basic_counts and a test
of cache_simulator using the scheduler's -cpu_schedule_file as-traced
mode.

Adds some dr$sim unit tests for cpuid to core mapping and error modes.

Issue: #5694