i#6499: AArch64 codec: mark instructions with governing predicate as predicated
This patch adds the DR_PRED_GOVERNING flag to instructions containing
any governing predicate registers as well as tagging the governing
register with DR_OPND_IS_GOVERNING.

To do this, it extends the codec functionalty such that operands
may have "flags" which allow us to add some special code generation
when they are encountered. In this case we ammend the instr decode/encode
functions to both tag the governing predicate and the instr itself.

issue: #i6499

Change-Id: If93aedb4036e7ec18b20bb6f56f4297b433737b9