i#5926 libdw: Add new interface dr_running_under_dynamorio()
Adds a new API routine dr_running_under_dynamorio() which is similar
to the application interface dr_app_running_under_dynamorio() but is
callable from a client.  The use case is code meant for us both in
managed mode under DR and in standalone library mode.  Sometimes the
behavior for the two differs significantly, such as whether certain
library routines are safe when linked statically.

Adds sanity checks to some existing tests.

The driving use case is in drsyms's support for libdw.  The actual use
of this will come in that libdw commit.

Issue: #5926