i#6538: Remove link lines and add find_library calls in drmemtrace_proj (#6539)
Removes the link lines from the drmemtrace_proj test code. This test
exercises the exported target code for drmemtrace and that code includes
all the needed static libraries on the link line. It is only missing
locating such libraries if not in standard places, so we add
find_library() calls for lz4 and snappy. We remove the warning about
zlib as it could be a false positive and seems superfluous.

Fixes #6538