i#6541 privlib: Add try/except around private library init/fini
Wraps private library init/fini calls in a try/except on Linux (as is
already done on Windows) and reports a crash as a SYSLOG_ERROR
pointing at the library's name.  This is much better than a weird
assert or mysterious secondary crash later.

Adds a check for shared_data==NULL in check_thread_in_last_vmwarea()
to avoid a crash seen there.

Issue: #6541