i#5740: Remove attach/detach flaky-list entries on a64
Now that we've enabled ptrace privileges on the a64 testing machine we
can remove the attach/detach tests from the flaky list as they no
longer hang.  The attach test passed 100x in a row for me so it
doesn't seem to be hitting flakes seen on other platforms like #6452.

Issue: #5740, #6558, #6127
Fixes #5740