i#6417 AMD 32-bit: Drastically reduce x86-32 tests run until AMD is fixed
We have lived with our x86-32 GA CI job being red from massive test
failures from the AMD switch for long enough.  The fix here may be too
drastic, but it's simple: we scale back to the tiny set of tests run
on ubuntu22.  I considered listing the failures on the ignore list,
but with timeouts included this job was just taking too long.

The plan is to have another Fixit and try to fix enough AMD failures
we can re-enable the full job.

Issue: #6417