i#6573: Remove the de-duplication logic of record filter
Remove the de-duplication logic for consecutive timestamp and cpu markers. Fix the problem that output entry count is not equal to the input entry count after run a null record filter test.
We added many more timestamps to drmemtrace traces beyond the original ones which were only at the start of a trace buffer. Additionally, zip-format traces have duplicated timestamps at the start of each chunk. This means that we sometimes have deliberate back-to-back timestamp;cpu pairs in existing traces, and we have updated trace consumers to be aware of this. This has caused us to revisit the removal of back-to-back pairs in the record filter and we have decided to just eliminate that feature: we're ok if a filter removing intervening records results in consecutive timestamp;cpuid pairs that were not consecutive before.
Fixes: #6573