i#6514 null SP to clone(), part 3: Handle null SP passed to clone
The kernel supports passing NULL for SP to clone: It means to use the SP
of the parent. Supporting this requires augmenting dynamorio_clone to
handle SP==NULL, and augmenting create_clone_record to handle SP==NULL.

Only X86 is supported in this pass. i#XYZ filed to handle the other arches.

Fixes #6514