i#6580: AArch64 codec.py: Remove variable reuse in function naming flow (#6595)
This patch refactors how the script gets hold of the name of generated
functions. Previously there was a pass in opndset_naming() that
overwrote the value of the opndset attribute. This was causing issues
with type checking because this changed the type of an opndset from a
tuple or string to always a string which was flagged as an error when
.split() was used on it.

As part of this change the fall-through code became completely dead code
and has been removed. It was obsoleted originally by the addition of +
masks for the operands.

It was also revealed by this change that several encodings in the codecs
were no longer used and represented less correct versions of other
encodings. They have been removed.

issue: #6580