i#6345: Delay branch instruction in rseq buffer after rollback. (#6346)
When rseq exits on the side, the code rolls back instructions in the
rseq buffer. If the last instruction being written from the rseq buffer
is a branch, it is not being delayed as in other cases. In the example
below, the branch, jbe, is followed by two pairs of timestamp+CPU
markers, and then the jmp instruction. The sequence triggers the "Branch
target not immediately after branch" invariant error.

1508045 ifetch 2 byte(s) @ 0x000000001153345d 76 56 jbe
$0x00000000115334b5 (taken)
     1508046    <marker: timestamp 13340822879775984>
     1508047    <marker: tid 812769 on core 25>
     1508048    <marker: timestamp 13340822879776016>
     1508049    <marker: tid 812769 on core 25>
1508050 ifetch 5 byte(s) @ 0x00000000115334b5 e9 61 ff ff ff jmp
$0x000000001153341b <<<<<

The change in this PR delays the branch instruction after rseq rollback,
so that it will be written after the timestamp+cpu markers.

Fixes #6345