i#5365: Infinite loop in linux.fib-conflict test when built with -O3
Remove linux.fib-conflict from list of tests to be builts with -03.

The cause of the infinite loop is hard to determine, being caused by the
linker script.

As this is not a DynamoRIO issue, and this test often fails at the
moment anyway, just build it without -O3 for now.

Issue: #5365