i#6635 core filter, part 3: Core-on-disk analysis support
Add core-sharded-on-disk analysis support

Adds a new filetype OFFLINE_FILE_TYPE_CORE_SHARDED set by
record_filter when core-sharded.

Has the scheduler read ahead to the filetype in every input, and
arrange to make it available at init time to output streams.

The analyzer looks for the filetype and sets the shard type as
core-sharded if set.  This is different from a dynamic-scheduling
core-sharded as we do not need a 1:1 shard:workers mapping.

Allows header records in reader_t::process_input_entry(), which are
now expected with core-sharded-on-disk traces.

Adds core-sharded reader support for seeing only portions of some
threads, where we have to relax some tid and pid checks.

Checks in 6 core-sharded-on-disk x64 threadsig traces for use in a
test of basic_counts analyzing them.  That analysis fails without all
of these framework changes.

The changes to the record filter to support generating those traces
in the first place will come in separately.

Issue: #6635