i#6643: Make interval_state_snapshot_t data members private
Marks the data members of the base interval_state_snapshot_t as private. This is
to formally disallow analysis tools that subclass this from modifying the data
members such as shard_id, interval_end_timestamp, etc., some of which are
essential to the interval merge logic implemented by the analyzer framework.

The analyzer framework implementation in analyzer_tmpl_t is marked as a friend
class so that it can still set those private data members.

Converted interval_state_snapshot_t from a struct to a class to follow our
style guide.

Issue: #6643, #6020