i#6666: Avoid hang when analyzer worker returns early
Adds a call to set_active(false) on the output stream when an analyzer
worker thread returns early due to an error.  This frees up the
current input and avoids a hang in the scheduler after the error is
printed.

Adds a unit test which hangs without this fix.

Fixes #6666