i#6635 core filter, part 4: record_scheduler_t inserted tid,pid
For record_scheduler_t, context switches cannot be observed as each
record has limited information and does not include the source.  We
insert a synthetic pair of tid and pid records on each switch to solve
this.

Updates the unit test to test this.

Updates the checked-in drmemtrace.threadsig-core-sharded.x64.tracedir/
trace files to include these records.  The files are also reduced to 4
cores to increase the switch count and renamed to avoid the tid in the
name.

Issue: #6635