i#6678 Rename aarchxx workflows, part 1: ci-aarchxx -> ci-aarchxx-cross (#6696)
There are currently two workflows name "ci-aarchxx" which is confusing.
The ci-aarch64-native workflow was renamed to ci-aarchxx in #6549 with
the goal of being consistent with other architecture workflows. However,
this introduced the ambiguous "ci-aarchxx" naming.

Fix this by renaming ci-aarchxx.yml -> ci-aarchxx-cross.yml, and
renaming ci-aarch64-native.yml -> ci-aarchxx.yml.
Furthermore, rename the aarchxx "cross" workflow to ci-aarchxx-cross.

This accomplishes three things:

1. Naming is consistent with the other workflow files: native workflows
are named "ci-$ARCH" and live in file "ci-$ARCH.yml".
2. All workflows have unique names.
3. The yml file and workflow have the same name.

In this part one we rename the cross ci-aarchxx to ci-aarchxx-cross, in
the workflow name and yml file name.
In part two to follow (#6697) we will leave the workflow name alone, but
still rename the file from ci-aarch64-native.yml to ci-aarchxx.yml. The
renaming is split into two to better preserve the history: Git can track
file renamings, but not when the original file is also replaced with
something different.

Issue #6678

Co-authored-by: Joshua Warburton <joshua.warburton@arm.com>