i#6712: Add record_scheduler_t support for replay-as-traced
Adds record_scheduler_t (for the record_filter tool) support for
replaying as-traced.  The key change here is having record_reader_t
increment the instruction ordinal on an encoding record instead of an
instruction record.  This avoids splitting encodings from instructions
when operating at instrution boundaries in the scheduler (during
replay, but this would also affect skipping for scheduler regions of
interest).

Adds a test of record_filter on the checked-in threadsig trace in
as-traced replay mode.

Fixes #6712