i#6713: Remove same-input-same-start cpu_schedule entries
Removes same-input-same-start cpu_schedule entries when reading the
schedule file in the scheduler.  The scheduler cannot emulate a switch
with no instructions in between in replay mode, and these same-start
entries cause problems.

Adds a scheduler unit test.

Tested on a larger app in as-traced mode where there are many of these
same-start entries and without this fix the scheduler advances too
early, causing problems.

Fixes #6713