i#6726 replay cpuid: Sort as-traced outputs by cpuid
Currently, in as-traced mode the output streams are assigned to the
cores in the as-traced schedule file in file order.  But that order is
essentially random, which scrambles key arrangements like which core
is on which socket.  Here we sort by the recorded cpuid to recreate
the same cpuid order as before.

Adds a unit test.  Also tested on larger cases with > 100 cores.

Issue: #6726