Fix failure on AARCH64.
In AARCH64 instr_get_arith_flags() may need to do additional decoding of instr.
If that happens, #dcontext_t ISA mode will be set to be the #intr_t ISA mode, which
is DR_ISA_SYNTHETIC.
This would trigger the decoding for DR_ISA_SYNTHETIC, which is not what we want.
So, we temporary set instr ISA mode to be the dcontex_t ISA mode, which contains
a real ISA mode (e.g., DR_ISA_ARM_A64).