i#3320 dr$sim bundle error: Mark multiproc test as ignore
The code_api|tool.drcachesim.multiproc test has failed several times
in release build on the sve runner post-merge.  Despite attempts to
figure out or theorize as to what is causing the failure, it remains a
mystery.  For now we add multiproc to the list of failures to ignore.

Issue: #3320