i#6728 sched time stats: Fix time accounting problems (#6784)
The schedule_stats tool was incorrectly counting instructions since the
last context switch leading up to an idle record as all idle time. It
was also failing to count the final idle record in a sequence. We fix
that here by simplifying and clarifying the time accounting.

While at it, a 3rd time bucket for wait time is added.

Adds a better test with a mock class with deterministic time to avoid
wall-clock timing problems in testing.

Fixes #6728