i#6793: Skip incorrect final interval due to missing thread exit
Skips generating the final interval for thread-shards that are
missing the thread exit entry. Such final intervals have
incorrect data (the cumulative instr count, instr count delta,
and end timestamp). Instead of generating incorrect data
we skip such a final interval and print an error message.

This issue is seen to happen with traces prior to the i#6444
fix that handles missing thread exit entries when tracing
ends in an rseq region. For such traces, the i#6444 fix
already added an analyzer workaround that ensures that
process_shard_exit is called despite the missing thread
exit event. Here we modify that workaround to skip
processing the final interval because the worker's trace
stream would have moved on to other shards (if there were
others) and therefore cannot provide the state (instr count,
timestamp as of then) required for interval generation.

Note that for the shard that was the last shard in the
worker's stream, we can still continue to generate the final
interval snapshot, because the worker's trace stream is still
good to generate state for that shard.

Adds modified version of unit tests to verify the workaround.

Fixes: #6793