i#6793: Skip incorrect final interval during trace analysis (#6794)
Skips generating the final interval snapshot for thread-shards that are
missing the thread exit entry, during trace interval analysis in the
drmemtrace analyzer framework. Such final intervals currently have
incorrect data (the cumulative instr count, instr count delta, and end
timestamp). Instead of generating incorrect data we skip such a final
interval and print a non-fatal error message.

This issue is seen on interval analysis of only the traces affected by
i#6444. That issue has already been fixed many months ago, and does not
affect newer traces. This fix is to allow trace interval analysis of
those older traces.

Due to i#6444, some threads were missing thread exit entries because
tracing ended in an rseq region. For such traces, the i#6444 fix already
added an analyzer workaround that ensures that process_shard_exit is
called despite the missing thread exit event. Here we modify that
workaround to skip processing the final interval when the workaround had
to be invoked; this is required because the worker's trace stream would
have moved on to other shards (if there were others assigned to the
worker) and therefore cannot anymore provide the state (instr count,
timestamp when the shard ended) required for generation of the final
interval snapshot for that shard.

Note that for the shard that was the last shard in the worker's stream,
we can still continue to generate the final interval snapshot, because
the worker's trace stream is still good to generate state for that
shard.

Adds modified version of existing interval analysis unit tests to verify
the workaround.

Also fixes a log which may not print the correct stream name if the
error was in a different shard than the current one (because of the same
reason we're seeing the primary issue).

Issue: #6444
Fixes: #6793